release v0.3
- Bugfixes in PAL and SHIM
- Improvement of migration scheme in SHIM, which largely speed up
  forking
- Upgrade glibc to 2.19
- FreeBSD port
- Support OpenJDK, python and R

+++ LibOS/shim/test/apps/lmbench/lmbench-2.5/src/bw_mmap_rd.c
@@ -25,7 +25,7 @@ char	*id = "$Id$\n";
 
 #define	TYPE	int
 #define	MINSZ	(sizeof(TYPE) * 128)
-#define	CHK(x)	if ((int)(x) == -1) { perror("x"); exit(1); }
+#define	CHK(x)	if ((long)(x) == -1) { perror("x"); exit(1); }
 
 void
 doit(register TYPE *p, register TYPE *lastone)
